Fix @id by moving GetCollection operation to the output Dto
