+++ composer.json
@@ -7,13 +7,13 @@
         "php": ">=8.2",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "api-platform/doctrine-orm": "^4.2.6",
-        "api-platform/symfony": "^4.2.6",
-        "doctrine/dbal": "^3.10.3",
+        "api-platform/doctrine-orm": "^4.2.10",
+        "api-platform/symfony": "^4.2.10",
+        "doctrine/dbal": "^3.10.4",
         "doctrine/doctrine-bundle": "^2.18.1",
-        "doctrine/orm": "^3.5.7",
+        "doctrine/orm": "^3.5.8",
         "nelmio/cors-bundle": "^2.6",
-        "phpdocumentor/reflection-docblock": "^5.6.4",
+        "phpdocumentor/reflection-docblock": "^5.6.5",
         "phpstan/phpdoc-parser": "^2.3",
         "symfony/asset": "7.4.*",
         "symfony/console": "7.4.*",
@@ -82,15 +82,15 @@
         }
     },
     "require-dev": {
-        "dama/doctrine-test-bundle": "^8.4",
-        "justinrainbow/json-schema": "^6.6.1",
-        "phpunit/phpunit": "^12.4.4",
+        "dama/doctrine-test-bundle": "^8.4.1",
+        "justinrainbow/json-schema": "^6.6.3",
+        "phpunit/phpunit": "^12.5.4",
         "symfony/browser-kit": "7.4.*",
         "symfony/css-selector": "7.4.*",
         "symfony/debug-bundle": "7.4.*",
         "symfony/http-client": "7.4.*",
-        "symfony/maker-bundle": "^1.65",
-        "symfony/monolog-bundle": "^3.10",
+        "symfony/maker-bundle": "^1.65.1",
+        "symfony/monolog-bundle": "^3.11.1",
         "symfony/stopwatch": "7.4.*",
         "symfony/web-profiler-bundle": "7.4.*"
     }
