+++ config/packages/api_platform.yaml
@@ -7,4 +7,4 @@ api_platform:
             vary: ['Content-Type', 'Authorization', 'Origin']
     mapping:
         paths:
-            - '%kernel.project_dir%/src/Api/Resource'
+            - '%kernel.project_dir%/src/ApiResource'
