--- src/Api/Dto/BookCollection.php
@@ -1,19 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Api\Dto;
-
-use App\Entity\Book as BookEntity;
-use Symfony\Component\ObjectMapper\Attribute\Map;
-
-#[Map(source: BookEntity::class)]
-final class BookCollection
-{
-    public int $id;
-
-    #[Map(source: 'title')]
-    public string $name;
-
-    public string $isbn;
-}
